# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Secure Gift Card Manager API
  version: 0.1.0
  description: >
    Initial API contract for the local MVP. The implementation must keep this
    file in sync with backend routes and QA requirement IDs.
servers:
  - url: http://localhost:3001
    description: Local API server
...
paths:
...
  /api/auth/setup:
    post:
      tags: [Auth]
      summary: Create the first account/user and wrapped DEK
      operationId: setupAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupRequest"
      responses:
        "201":
          description: Setup complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Unlock encrypted data
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/cards:
    get:
      tags: [Cards]
      summary: List cards
      operationId: listCards
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CardStatus"
        - name: brand
          in: query
          schema:
            type: string
        - name: cardNumber
          in: query
          description: Exact full card-number search. Backend normalizes and blind-indexes input.
          schema:
            type: string
      responses:
        "200":
          description: Cards page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Cards]
      summary: Create one or more cards
      operationId: createCards
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardsRequest"
      responses:
        "201":
          description: Cards created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/deals:
    get:
      tags: [Deals]
      summary: List deals
      operationId: listDeals
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: includeArchived
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Deals page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Deals]
      summary: Create a deal, optionally with cards
      operationId: createDeal
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDealRequest"
      responses:
        "201":
          description: Deal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/settings/backup:
    get:
      tags: [Settings]
      summary: Get backup settings and backup status
      operationId: getBackupSettings
      responses:
        "200":
          description: Backup settings
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ObjectResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/BackupSettings"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [Settings]
      summary: Update backup settings
      operationId: updateBackupSettings
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupSettingsUpdateRequest"
      responses:
        "200":
          description: Backup settings updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ObjectResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/BackupSettings"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/users:
    get:
      tags: [Users]
      summary: List account users
      operationId: listUsers
      responses:
        "200":
          description: Account users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Users]
      summary: Create an account user
      operationId: createUser
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ObjectResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/audit:
    get:
      tags: [Audit]
      summary: List audit events
      operationId: listAudit
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Audit page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditListResponse"
...
components:
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 128
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 16
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Not authenticated or locked
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated request failed CSRF, origin, or authorization checks
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: State conflict or duplicate
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    AuthStatusResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [setupComplete, sessionValid, dekLoaded]
          properties:
            setupComplete:
              type: boolean
            sessionValid:
              type: boolean
            dekLoaded:
              type: boolean
            user:
              $ref: "#/components/schemas/AuthUser"
            csrfToken:
              type: string
    AuthUser:
      type: object
      required: [id, accountId, role]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        role:
          type: string
          enum: [owner, admin, operator, viewer]
        email:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
...
    SetupRequest:
      type: object
      required: [unlockSecret]
      properties:
        unlockSecret:
          type: string
          minLength: 8
        displayName:
          type: string
        email:
          type: string
          format: email
    LoginRequest:
      type: object
      required: [unlockSecret]
      properties:
        email:
          type: string
          format: email
          description: Required when more than one active user exists.
        unlockSecret:
          type: string
...
    CardStatus:
      type: string
      enum: [available, reserved, in_use, sold, used_up, void]
    CardType:
      type: string
      enum: [merchant, prepaid]
    Network:
      type: string
      enum: [visa, mastercard, amex, discover, other]
    Card:
      type: object
      required:
        - id
        - accountId
        - brand
        - cardType
        - faceValueCents
        - remainingBalanceCents
        - purchaseCostCents
        - status
        - rowVersion
      properties:
        id:
          type: integer
        accountId:
          type: integer
        dealId:
          type: integer
          nullable: true
        brand:
          type: string
        cardType:
          $ref: "#/components/schemas/CardType"
        network:
          $ref: "#/components/schemas/Network"
        faceValueCents:
          type: integer
          minimum: 1
        remainingBalanceCents:
          type: integer
          minimum: 0
        purchaseCostCents:
          type: integer
          minimum: 0
        cardNumberLast4:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        expirationDate:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/CardStatus"
        reservedFor:
          type: string
          nullable: true
        reservedUntil:
          type: string
          nullable: true
        reservedNotes:
          type: string
          nullable: true
        latestSalePriceCents:
          type: integer
          nullable: true
        rowVersion:
          type: integer
    CardInput:
      type: object
      required: [brand, cardType, faceValueCents]
      properties:
        brand:
          type: string
        cardType:
          $ref: "#/components/schemas/CardType"
        network:
          $ref: "#/components/schemas/Network"
        faceValueCents:
          type: integer
          minimum: 1
        purchaseCostCents:
          type: integer
          minimum: 0
        cardNumber:
          type: string
        pin:
          type: string
        billingZip:
          type: string
        expirationDate:
          type: string
        notes:
          type: string
    CreateCardsRequest:
      type: object
      required: [cards]
      properties:
        cards:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CardInput"
...
    Deal:
      type: object
      required: [id, accountId, name, rowVersion]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        name:
          type: string
        source:
          type: string
          nullable: true
        purchaseDate:
          type: string
          nullable: true
        inputTotalCostCents:
          type: integer
          nullable: true
          minimum: 0
        notes:
          type: string
          nullable: true
        archivedAt:
          type: string
          nullable: true
        rowVersion:
          type: integer
    CreateDealRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        source:
          type: string
        purchaseDate:
          type: string
        totalCostCents:
          type: integer
          minimum: 0
        notes:
          type: string
        cards:
          type: array
          items:
            $ref: "#/components/schemas/CardInput"
...
    AuditEvent:
      type: object
      required: [id, accountId, entityType, action, timestamp]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        entityType:
          type: string
        entityId:
          type: integer
          nullable: true
        action:
          type: string
        timestamp:
          type: string
...
    CardListResponse:
      allOf:
        - $ref: "#/components/schemas/ListEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Card"
...
    DealDetailResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [deal, cards]
          properties:
            deal:
              $ref: "#/components/schemas/Deal"
            cards:
              type: array
              items:
                $ref: "#/components/schemas/Card"
    DealListResponse:
      allOf:
        - $ref: "#/components/schemas/ListEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Deal"
    AuditListResponse:
      allOf:
        - $ref: "#/components/schemas/ListEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/AuditEvent"
...
    BackupSettings:
      type: object
      required:
        - allowPlaintextExport
        - plaintextExportPolicyLocked
        - backupReminderDays
        - backupReminderDue
        - lastBackupAt
        - nextBackupDueAt
        - lastPlaintextExportAt
        - lastEncryptedExportAt
        - lastRawDatabaseExportAt
      properties:
        allowPlaintextExport:
          type: boolean
          description: Effective plaintext export availability after account settings and deployment policy are applied.
        plaintextExportPolicyLocked:
          type: boolean
          description: True when `GC_PLAINTEXT_EXPORT_ENABLED=false` prevents plaintext JSON export from being enabled.
        backupReminderDays:
          type: integer
          minimum: 0
          maximum: 365
          description: 0 disables the reminder.
        backupReminderDue:
          type: boolean
        lastBackupAt:
          type: string
          format: date-time
          nullable: true
        nextBackupDueAt:
          type: string
          format: date-time
          nullable: true
        lastPlaintextExportAt:
          type: string
          format: date-time
          nullable: true
        lastEncryptedExportAt:
          type: string
          format: date-time
          nullable: true
        lastRawDatabaseExportAt:
          type: string
          format: date-time
          nullable: true
    BackupSettingsUpdateRequest:
      type: object
      required: [unlockSecret, allowPlaintextExport, backupReminderDays]
      properties:
        unlockSecret:
          type: string
        allowPlaintextExport:
          type: boolean
        backupReminderDays:
          type: integer
          minimum: 0
          maximum: 365
    User:
      type: object
      required: [id, accountId, displayName, role, createdAt, updatedAt]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        email:
          type: string
          nullable: true
        displayName:
          type: string
        role:
          type: string
          enum: [owner, admin, operator, viewer]
        disabledAt:
          type: string
          format: date-time
          nullable: true
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
    UserCreateRequest:
      type: object
      required: [currentUnlockSecret, email, displayName, role, unlockSecret]
      properties:
        currentUnlockSecret:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          type: string
          enum: [admin, operator, viewer]
        unlockSecret:
          type: string
          minLength: 8
          description: Temporary unlock secret for the new user.
...
    ListEnvelope:
      type: object
      required: [data, page]
      properties:
        data:
          type: array
        page:
          type: object
          required: [limit, offset, total, hasMore]
          properties:
            limit:
              type: integer
            offset:
              type: integer
            total:
              type: integer
            hasMore:
              type: boolean
    ObjectResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            fieldErrors:
              type: array
              items:
                type: object
                required: [field, code, message]
                properties:
                  field:
                    type: string
                  code:
                    type: string
                  message:
                    type: string